+++ apps/condo/domains/resident/tasks/discoverServiceConsumers.js
@@ -1,8 +1,7 @@
 const { getLogger } = require('@open-condo/keystone/logging')
 const { getSchemaCtx } = require('@open-condo/keystone/schema')
-const { createTask } = require('@open-condo/keystone/tasks')
 
-const { discoverServiceConsumers } = require('@condo/domains/resident/utils/serverSchema')
+const { discoverServiceConsumers: discoverServiceConsumersMutation } = require('@condo/domains/resident/utils/serverSchema')
 
 const DV_SENDER = { dv: 1, sender: { dv: 1, fingerprint: 'discoverServiceConsumersTask' } }
 const logger = getLogger('discoverServiceConsumers')
@@ -11,11 +10,11 @@ const logger = getLogger('discoverServiceConsumers')
  * @param {DiscoverServiceConsumersInput} data
  * @returns {Promise<void>}
  */
-async function discoverServiceConsumersTask (data) {
+async function discoverServiceConsumers (data) {
     const { keystone: context } = getSchemaCtx('Resident')
 
     try {
-        const result = await discoverServiceConsumers(context, { ...data, ...DV_SENDER })
+        const result = await discoverServiceConsumersMutation(context, { ...data, ...DV_SENDER })
         logger.info({ msg: 'Done', result })
     } catch (err) {
         logger.error({ msg: 'Error', err })
@@ -23,5 +22,5 @@ async function discoverServiceConsumersTask (data) {
 }
 
 module.exports = {
-    discoverServiceConsumersTask: createTask('discoverServiceConsumers', discoverServiceConsumersTask, { priority: 10 }),
+    discoverServiceConsumers,
 }
