+++ apps/condo/domains/resident/tasks/index.js
@@ -1,19 +1,32 @@
-const { discoverServiceConsumersTask } = require('./discoverServiceConsumers.task')
-const { discoverServiceConsumersCronTask } = require('./discoverServiceConsumersCron.task')
+const { createTask, createCronTask, removeCronTask  } = require('@open-condo/keystone/tasks')
+
+const { actualizeResidentToPropertyAndOrganizationConnections } = require('./actualizeResidentToPropertyAndOrganizationConnections')
+const { actualizeTicketToResidentUserConnections } = require('./actualizeTicketToResidentUserConnections')
+const { discoverServiceConsumers } = require('./discoverServiceConsumers')
+const { discoverServiceConsumersLastDate } = require('./discoverServiceConsumersLastDate')
+const { notifyResidentsOnPayday } = require('./notifyResidentsOnPayday')
 const {
-    manageResidentToPropertyAndOrganizationConnections,
-    manageResidentToTicketClientConnections,
-} = require('./residentTicket.task')
-const {
-    sendBillingReceiptNotificationsWorkDaysTask,
-    sendBillingReceiptNotificationsWeekendsTask,
-} = require('./sendBillingReceiptNotifications.task')
+    sendBillingReceiptNotificationsWorkDays,
+    sendBillingReceiptNotificationsWeekends,
+} = require('./sendBillingReceiptNotifications')
+const { sendMessageToResidentScopes } = require('./sendMessageToResidentScopes')
+
+// Renamed to "discoverServiceConsumersLastDate"
+removeCronTask('discoverServiceConsumersCronTask', '13 * * * *')
+// Renamed to "sendBillingReceiptNotificationsWorkDays"
+removeCronTask('sendBillingReceiptNotificationsWorkDaysTask', '17 7-12 * * 1-5')
+// Renamed to "sendBillingReceiptNotificationsWeekends"
+removeCronTask('sendBillingReceiptNotificationsWeekendsTask', '17 9-12 * * 0,6')
+
+// PODS work according to UTC, so time in cron-tasks should also be according to UTC too.
 
 module.exports = {
-    sendBillingReceiptNotificationsWorkDaysTask,
-    sendBillingReceiptNotificationsWeekendsTask,
-    manageResidentToPropertyAndOrganizationConnections,
-    manageResidentToTicketClientConnections,
-    discoverServiceConsumersCronTask,
-    discoverServiceConsumersTask,
+    sendBillingReceiptNotificationsWorkDaysCronTask: createCronTask('sendBillingReceiptNotificationsWorkDays', '17 7-12 * * 1-5', sendBillingReceiptNotificationsWorkDays),
+    sendBillingReceiptNotificationsWeekendsCronTask: createCronTask('sendBillingReceiptNotificationsWeekends', '17 9-12 * * 0,6', sendBillingReceiptNotificationsWeekends),
+    actualizeResidentToPropertyAndOrganizationConnectionsTask: createTask('actualizeResidentToPropertyAndOrganizationConnections', actualizeResidentToPropertyAndOrganizationConnections),
+    actualizeTicketToResidentUserConnectionsTask: createTask('actualizeTicketToResidentUserConnections', actualizeTicketToResidentUserConnections),
+    discoverServiceConsumersLastDateCronTask: createCronTask('discoverServiceConsumersLastDate', '13 * * * *', discoverServiceConsumersLastDate, { priority: 10 }),
+    discoverServiceConsumersTask: createTask('discoverServiceConsumers', discoverServiceConsumers, { priority: 10 }),
+    notifyResidentsOnPayDayCronTask: createCronTask('notifyResidentsOnPayday', '0 13 * * *', notifyResidentsOnPayday),
+    sendMessageToResidentScopesTask: createTask('sendMessageToResidentScopes', sendMessageToResidentScopes),
 }
