+++ apps/condo/domains/resident/tasks/notifyResidentsOnPaydayTask.js
@@ -1,39 +1,25 @@
-const dayjs = require('dayjs')
+/**
+ * @jest-environment node
+ */
 
-const { featureToggleManager } = require('@open-condo/featureflags/featureToggleManager')
-const { getLogger } = require('@open-condo/keystone/logging')
-const { createCronTask } = require('@open-condo/keystone/tasks')
+const index = require('@app/condo/index')
 
-const { SEND_BILLING_RECEIPTS_ON_PAYDAY_REMINDER_TASK } = require('@condo/domains/common/constants/featureflags')
+const { setFakeClientMode, setAllFeatureFlags } = require('@open-condo/keystone/test.utils')
 
-const { notifyResidentsOnPayday } = require('./notifyResidentsOnPayday')
+const notifyResidentsOnPaydayTask = require('./notifyResidentsOnPayday')
 
 
-const logger = getLogger('meter/sendSubmitMeterReadingsPushNotifications')
+describe('Meter verification notification', () => {
+    setFakeClientMode(index)
+    describe('feature flag', () => {
+        setAllFeatureFlags(true)
+        it('checks for proper result on enabled', async () => {
+            expect(await notifyResidentsOnPaydayTask).toBeUndefined()
+        })
 
-module.exports = createCronTask('notifyResidentsOnPayday', '0 13 * * *', async (context = null) => {
-    const isFeatureEnabled = await featureToggleManager.isFeatureEnabled(context, SEND_BILLING_RECEIPTS_ON_PAYDAY_REMINDER_TASK)
-
-    if (!isFeatureEnabled) {
-        logger.info(`Skip notifyResidentsOnPayday task due to disabled GrowthBook feature flag [${SEND_BILLING_RECEIPTS_ON_PAYDAY_REMINDER_TASK}]`)
-
-        return 'disabled'
-    }
-
-    const today = dayjs()
-    const targetDayOfSendingThePush = dayjs().set('date', 20).day()
-    const isTargetDayWeekend = targetDayOfSendingThePush === 0 || targetDayOfSendingThePush === 6
-    const isTargetDayWeekendAndTodayIsFriday = isTargetDayWeekend && Number(today.format('DD')) >= 18 && Number(today.format('DD')) < 20 && today.day() === 5
-    const isTargetDayWeekday = !isTargetDayWeekend && today.format('DD') === '20'
-
-    /**
-     * Pay date is:
-     * either a 20-th day workday of month
-     * or Friday coming before weekend 20-th day
-     */
-    if (isTargetDayWeekendAndTodayIsFriday || isTargetDayWeekday) {
-        return await notifyResidentsOnPayday()
-    } else {
-        logger.info({ msg: 'Push should be sent only on weekdays.' })
-    }
+        it('checks for proper result on disabled', async () => {
+            setAllFeatureFlags(false)
+            expect(await notifyResidentsOnPaydayTask).toEqual('disabled')
+        })
+    })
 })
