+++ apps/condo/domains/resident/tasks/sendBillingReceiptNotifications.js
@@ -1,0 +1,14 @@
+const { sendBillingReceiptNotifications } = require('./helpers')
+
+async function sendBillingReceiptNotificationsWorkDays () {
+    return await sendBillingReceiptNotifications()
+}
+
+async function sendBillingReceiptNotificationsWeekends () {
+    return await sendBillingReceiptNotifications()
+}
+
+module.exports = {
+    sendBillingReceiptNotificationsWorkDays,
+    sendBillingReceiptNotificationsWeekends,
+}
